requirements.txt
streamlit
python-pptx
PyGithub